pyproject.toml
[tool.poetry]
name = "familysearchai"
version = "0.1.0"
description = ""
authors = ["timytapilla <ajdnmkem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
mydependency = "^0.0.0"
requests = "^2.32.3"
markdownify = "^0.13.1"
numpy = "1.26.4"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
unidecode = "^1.3.8"
wordcloud = "^1.9.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
textblob = "^0.18.0.post0"
scikit-learn = "^1.5.2"
networkx = "^3.3"
llama-index = "^0.11.14"
pinecone-client = "^5.0.1"
pinecone = "^5.3.1"
llama-index-core = "^0.11.14"
llama-index-vector-stores-pinecone = "^0.2.1"
python-frontmatter = "^1.1.0"
gradio = "^4.44.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
